Call RmShutdown() with proper action flags
[1] states that the "lActionFlags" parameter has to be "One or more
RM_SHUTDOWN_TYPE options" that "can be combined by an OR operator". Thus,
passing 0 is invalid, valid options are combinations of "RmForceShutdown"
and "RmShutdownOnlyRegistered". As passing 0 surprisingly does not cause
an error, but results in all applications to be shut down, no matter
whether they can be restarted or not, make this undocumented behavior
explicit by passing "RmForceShutdown or RmShutdownOnlyRegistered".

[1]
http://msdn.microsoft.com/en-us/library/windows/desktop/aa373667%28v=vs.85%29.aspx