include search order for ISPP does not match documentation
The behavior I see is that it's <filename> that searches only the current include path, and "filename" that also searches the directory where the current file resides. The observed behavior also matches what I would have expected based on the rest of ISPP being somewhat based on the C preprocessor and it's <system> and "local" include syntax.

The code that seems to implement the rule is clearly setting IncludePathOnly := True when the <> delimiters *are* used.

So the documentation just appears to be backwards.