Uninstaller: copy second phase executable into a safe location
Regular users' `%TMP%` directory might be a safe location because other
users cannot write into it. But SYSTEM's `%TMP%` is set to
`C:\Windows\Temp` by default, which is typically world-writable.

This opens the uninstaller up to DLL hijacking when run under the SYSTEM
user (note: this is different from merely running the uninstaller as an
admin user with elevated privileges).

Let's just do the same as we do in the installer: instead of copying or
extracting the executable into `%TMP%` directly, let's just create a
subdirectory and copy/extract into that. With this, we can guarantee
that no malicious `.dll` has been placed there.

Note: Since the second phase executable cannot delete itself, we need to
use a non-random name (only admins can run `RestartReplace()`
successfully, that function will fail if other users call it, therefore
we want at least that the temporary directory is reused the next time
an uninstaller is run). We still will want to ensure that the directory
has been newly created; To that end, we first try to delete any file or
diretory that is in the way, and only then create it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>