Avoid creating temporary directories with too-loose permissions
By default, when the SYSTEM account creates a directory in
`C:\Windows\Temp` permissions are inherited from that parent directory.

But those permissions are too loose! Every authenticated user could have
a background process that listens to filesystem events, quickly sneaking
in a `.dll` into the directory that Inno Setup just created.

Let's use a very restrictive DACL to prevent precisely that: only the
owner gets Full Access, and files/directories created within the new
directory will inherit these ACLs.

Note: To be able to use the `CreateSafeDirectory()` function that was
just added to `SecurityFunc.pas`, we now have to compile `SetupLdr.dpr`
with access to `Forms` to avoid a compile error.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>