Explicitly allow the current user full permission on temporary directories
When creating directories in elevated mode with ACLs restricted via
"Owner Rights", un-elevated access is actually forbidden.

This is a problem when running the installer in the IDE: In that case,
the IDE (which runs non-elevated) cannot clean up the temporary
directories.

To help that, let's add explicit permissions for the current user.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>