Fix evaluation of #elif <expr> inside false context
e.g.
#if 0
#if 1
#elif NOEXIST
#endif
#endif

Should not evalute NOEXIST (as it would raise an "undefined identifier" error)

#elif *should* evaluate when the Stack.Last.BlockState is false
(due to the preceding #if to which it offers an alternative)

But "or not Stack.Last.BlockState" not only reverses the innermost scope
being false, it also overrules any outer scope that was being skipped.
What we want is just to just skip even checking the Last BlockState,
while honoring its parents