README: describe how to set up code-signing in the CI runs
The Continuous Integration ("CI") definition provides a very convenient
way to build InnoSetup.

There is just that one downside that the build artifacts (including the
InnoSetup compiler) are not code-signed.

However, as part of the security work I did leading up to
https://github.com/git-for-windows/git/security/advisories/GHSA-gf48-x3vr-j5c3
I really needed a code-signed InnoSetup compiler, and added support for
producing precisely such a CI build artifact in 9572aae (ci: use
code-sign certificate if available, 2022-03-16).

This support was undocumented, though. Let's address that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>