Changed the self coded user:password filtering by Delphi internal functions and repaired user and password encoding for non us characters
I found out, that Delphi already has functions to get username and password from URIs. I therefore changed my function using these internal delphi functions to have a better code.

While testing, I found out, that at least with 10.4.2 username:password URLs will only work, if the username does not include non us-ascii characters and the password does not include spaces and non-us-characters. It looks like the URL-encoding takes place too early and therefore the username and passwords are sent url-encoded, too. So I have changed the way the request is formed. I strip username and password completely from the url and create an authorization header. In this way, the system works great (using Delphi 10.4.2 to compile). I have tested usernames and passwords with umlauts and spaces. I have added a note in the help, that usernames and spaces MUST NOT be url-encoded (this is no change in behaviour, but this is the way it works)