Vista hack: help keeping the expected block size up to date
Ever since introducing the hack to modify the embedded manifest, we have
a relatively opaque error message whenever the DLL protections in the
manifest are modified because we have a constant that needs to be
adjusted accordingly.

Let's make it easier to adjust that constant by letting the error
message state explicitly what the correct value should be.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>