Added a SetDownloadCredentials procedure
Since the Delphi frame work (at least 10.4.x) has limitations using urls with username and passwords (can not work with url encoded usernames and passwords and has parsing issues when not url encoded) I've added a SetDownloadCredentials procedure as workaround. The special urls can still be used and the limitations (not able to use characters like ? in the password) are unchanged; the logging will still not show the passwords. But using the SetDownloadCredential procedure, the username and password is used directly to form the basicauth header and therefore all parsing and encoding problems are gone.

In conclusion: unchanged scripts will work without the need of any change. The logging will not show the password. Some enhancements are done, so most characters are allowed in usernames and passwords. But using the special urls is deprecated.
Therefore using the new procedure will solve all encoding problems, but this will need a small script change; so only new/changed scripts will benefit.