Added an AddEx procedure to the TDownloadWizardPage
As requested, I added the option to use UserName & Password to the DownloadWizardPage.

The Add procedure is unchanged, but I added the AddEx procedure, so one can use different credentials for each download. In this way the change is not breaking existing installation scripts

The changes are straight forward, I enhanced the TDownloadFile object and used the new procedure SetDownloadCredentials before calling the download function. The help was updated, too