Draw menu as inactive when app is inactive. Uses GetForegroundWindow and works exactly like light menu's without having to listen to extra events. Looking at Delphi's Active doesnt work because it lags.
Opening a dialog within the app doesn't trigger the inactive look, unlike light menu's: there doesn't seem to be a draw event happening on opening a dialog (only on closing). In fact, it seems like redrawing of the menu is entirely disabled while the dialog is open. I suppose it uses a buffer.

(Edit, comment got lost:) BUT... On Windows 7 it doesn't buffer like that and dragging the dialog over the menu makes it drawn half active half inactive. So not done.