Simplify Windows version/architecture logging.
- Drop "(NT platform: Yes)"

- Combine "64-bit Windows" and "Processor architecture" into "Windows architecture"

I never liked the term "processor architecture" because although it's the term used by the GetSystemInfo API, it sounds like it's querying the physical processor. But it's not; it's reporting the architecture of the Windows installation, which may be different from the processor's native architecture (e.g., 32-bit x86 Windows can run on an x64 processor).

(Not going to rename ProcessorArchitecture etc. as that's way too invasive.)