Update comment.
BTW: there's also a REGEX_MULTILINE define which is off by default but when building with VS turning it on it does nothing. This is not clear from the Scintilla documentation but can be seen in Cxx11RegexFindText: it checks that _MSC_VER isn't defined but it is. Removing this check gives a compiler error that std::regex::multiline doesn't exist. Also see https://sourceforge.net/p/scintilla/bugs/2338/