whatsnew: Replace colors.
The previous colors seemed to have a purple tint to them when viewing on a newer monitor. Don't know if the problem is with the colors or the monitor.

Switching to the same blue shade (hue ~206, 100% saturation) used in the Windows 10/11 UI (e.g., on selected text, and the blue button in Calculator).

The new color for unvisited links is also nearly identical to the Firefox default, which appears to be hsl(210 100 40).

Visited links are somewhat lighter, like before. Not using a lighter-than-visited color for active links anymore (the difference was barely noticeable).