IDE: Use proper owner on message boxes.
AppMessageBox wasn't compatible with MainFormOnTaskBar=True; it displayed every message box with no owner.

Also found that the VCL doesn't set the owner correctly on dialogs displayed when the app isn't in the foreground. Workaround implemented.

Left the IsIconic call (from #164) in for now, though I don't think it's needed anymore in the IDE as long as message boxes use the main form as their owner. It may still be needed by Setup; haven't yet checked.