Unregister .iss association from both HKCU and HKLM, and remove all the conditional deletion code.
Before the "Open with" menu existed, if an app wanted to add itself to an existing association's context menu without taking over the association completely, it had to modify the ProgID key of whatever app owned the association. To allow for that, we previously were very careful to only delete keys/values under InnoSetupScriptFile that we created.

But there's no longer any need for that. If an app wants an "Open with" item, it should register its own ProgID and add a value under ".iss\OpenWithProgids" linking to it.