Doc: Make topic heading sticky.
This uses "position: sticky" so it doesn't work on IE 11 / HTML Help.

This *could* be IE-compatible if it instead used "position: fixed", but IE also doesn't support "scroll-padding" so there would be a problem with anchor targets being hidden underneath the header (needing hacks to fix).

Another approach is to make only .topicbody scrollable, not the whole page (like contents.htm). It works well on IE 11 (using flexbox, not grid) but seems to create an accessibility problem: if you focus the topic frame, pressing Up and Down does nothing until you tab to a link inside .topicbody, or tab to .topicbody itself (which requires tabindex="0" in IE 11).