Add TDownloadWizardPage support class functions AddWithISSigVerify and AddExWithISSigVerify. Todo: doc & update CodeDownloadFiles2.iss. Also doc something to prevent people from doing issigverify twice on a downloaded archive.
Could not use array of string in these new functions and had to use TStringList. For consistency changed support function ISSigVerify to use TStringList as well.

Also, didn't need this in the end but kept it anyway: make TDownloadWizardPage's and TExtractionWizardPage's Add(Ex) a function returning the index of the added item, like all other adds.