Call Abort immediately when OnExtractionProgress returns False.
Previously, if extraction completed while OnExtractionProgress was running
or soon after, then the function would succeed and not raise EAbort
despite the False result. This was confusing because it made it look like
returning False was having no effect. It would also be problematic if the
user's code performed some sort of teardown action prior to returning
False, but then no exception was raised and execution continued normally.