Support implicit conversion between Integer64 and Int64.
This allows the two types to be used interchangeably in assignment
statements and function parameters.

I don't intend to implement more operators (e.g., addition or comparison)
because the ultimate goal is to replace Integer64 usage with Int64 (or
UInt64 in certain cases). This just allows the migration to be done more
incrementally.