Allow VCL to style buttons. The VCL style actually looks better and this also makes it work on old versions.
The current state is now:

- Compil32's dialogs are all dark now, regardless of Windows dark mode.

- Windows' dialogs are also all dark now, except for TaskDialog. The list of dialogs is: Open File, Save File, Browse For Folder, Find, Replace, and Choose Font. TaskDialog is todo although I don't know how.

- Some are only partially dark if Compil32 is in dark mode but Windows is not. This seems acceptable to me. So not a todo.

- Some of dialogs still have some non-dark controls, even if Windows is in dark mode.

  - This applies to Windows' dialogs: Browse For Folder, Find, Replace, and Choose Font. However, it's not a large amount on non-darkness and seems acceptable to me. So not a todo.

  - It also applies to TNewChecklistbox in the wizard: bits of it are dark but not everything. Looking at that is a todo.

  - The images in the wizard are a todo.

All of the above applies to Windows 11. Todo count: 3.

On Windows 7 and other old versions the following things are not dark: Open File, Save File, the title bar, and the main form's scrollbars. The rest looks same as on Windows 11.

I suppose this means that when TNewChecklistbox is done and TNewProgressBar also, it has become possible to add style support to Setup as well.