Add VCL Styles support to TNewCheckListBox checkmarks.
I think it could also be improved by getting the style to draw the text, but don't need that at the moment, it looks fine as is.

I suppose it would be about getting the font name/size/style from the VCL style if seFont is on but I tried to edit the VCL style's font face for various elements in the .vsf and nothing is happening at run time (but if I change the font color of same elements then that does show up). Not sure what's up with that. It loads the font names/sizes/styles from the .vsf in Vl.StyleAPI's TSeStyleFonts.LoadFromStream. When browsing to the VCL code for various components I only see it used the color.