Cleanup: PerformInstall was basically impossible to navigate and read because of its many local functions, with many more local functions inside those.
Fix this by making the top level local functions regular functions. Then move out all those which dont access any install state such as UninstLog to a new unit. So for example ProcessInstallDeleteEntries was moved but RecordUninstallDeleteEntries was not.

The remaining top level functions in Setup.Install now all use install state even though their parameter list doesnt show this. This could be further cleaned up for better readability by using parameters but not doing this atm because unlike the above that would be refactoring with a risk of bugs.

So this commit doesn't actually change anything and doesnt contain any refactoring.