Introduce a way to tell the compiler it should apply dark colors/images/icons for Setup. Would be True if a dark VCL style was used (which you currently can't have yet). Doesn't actually make it use dark images/icons yet.
Note that even if one could freely choose a .vsf (VSL Style file), it would still need to know with a setting like this if it's a dark one or not. Because not all of them are dark.

This stuff isn't in its final form yet since ultimately I would like Setup to be able to choose light or dark mode based on what Windows' current mode is at runtime (but only when specified by the script). This means Setup would need both sets of colors/images/icons.