Add second Setup project. -Setup.dpr includes Components\Themes, SetupCustomStyle.dpr includes Vcl.Themes and Vcl.Styles. -Additionally SetupCustomStyle.dpr defines VCLSTYLES and includes a VCL Styles file (.vsf). Setup.dproj does neither.
Doesnt mean there are a ton of VCLSTYLES defines everywhere, quite the opposite: only 3 components check for it, to omit their style hooks. This is achieved using the dummy Vcl.Themes unit in Components\Themes.pas.

Not yet integrated into the build etc. Also no pfSetupCustomStyleE32 yet.

The gains are:
-The Setup project and Setup.e32 is still totally free of any Vcl.Themes and Vcl.Styles code, and therefore free of any potentional quirks it might have.
-Using Setup.e32 instead of SetupCustomStyle.e32 saves 116 kb in the end, not counting the style of the .vsf style itself. Including the style the difference is about 200 kb in the end.