Randomize uninstall temp directory names (no more iu-14D2N.tmp)
This introduces a new cleanup routine that runs when the uninstaller is
started, which checks for and deletes unused temp directories. As before,
there should typically only be 1 or 2 temp directories left behind (2
when a non-admin uninstaller runs after an admin uninstaller), but there
can be more if uninstallers are run concurrently or if the Temp directory
is shared by multiple users (C:\Windows\Temp).

The cleanup does require "List" access on the Temp directory, which
unprivileged users lack on C:\Windows\Temp. But only the SYSTEM account
is known to use C:\Windows\Temp (and only on outdated systems that don't
have GetTempPath2); everyone else (including service accounts) should be
using the Temp directory within their profile.