Enable RedirectionGuard on non-elevated Setup process (spawn server) also.
Mainly just for consistency between "runascurrentuser" and
"runasoriginaluser". There's not much security value in doing this,
because if an installer is running an EXE in a path that is writable by
non-admin users, there's a bigger problem: the non-admin users probably
have the ability to replace the EXE. No need to create redirects.