Remove the ArchitecturesAllowed versus x64 check, because it doesn't work after all, or at least not as you would expect.
'ArchitecturesAllowed=not x86' doesnt work because thats true when running Setup.e32 on a x64 processor (remember that 'x86' is actually 'x86os'). 'ArchitecturesAllowed=not x86compatible' also doesnt work because thats false when running Setup.e64 on a x64 processor.

So we just leave it to Windows to complain it cant run the exe.

Partially reverts 55c8d6ab.