Add x64 support to ASMInline and CreateCallback. Includes partial cleanup of x86 support in ASMInline, with no actual changes.
This also adds NativeInt to ROPS, but not yet in its final place, should be defined by ROPS itself, so it can use it in uPSC*. Also no NativeUInt yet.

Partially updates CodeDll.iss for x64, but some NativeInt should be NativeUInt (or even better: the actual specific type), and also only the callback stuff is updated and not anything else yet.

Having a NativeInt does introduce a new problem? 64-bit Uninstall can no longer assume it can safely run [Code] stored by 32-bit Uninstall, and vice versa? So when getting the utCode to run it should not only check the Setup version but also the bitness?